Sunflow Experiment Data

+++ tools/bootloader/sys.h
@@ -151,7 +151,7 @@ EXTERNAL int ProfileInit();
 EXTERNAL void register_thread_stat();
 EXTERNAL void init_log_queue(int,int);
 EXTERNAL void print_logs();
-EXTERNAL void add_log_entry(double*,int,long,int);
+EXTERNAL void add_log_entry(double*,int,long long,int);
 EXTERNAL int add_method_entry(char*,char*);
 EXTERNAL void end_iteration();
